refactor(event-store): extract the durable log's read path into a shared package
Move resolveEventStorePaths, loadGenesis, loadEventLog,
loadFoldedReview, readOptional, assertLogFullyLoaded and the
quarantine helpers out of apps/tui/src/event-store.ts (596 -> 433
lines) into a new private @numisma/event-store package, so both the
TUI and the web push can fold the same durable log instead of each
growing its own copy.

This is a move, not a rewrite: the tests that cover this code move
with it unchanged, and the suite lands exactly at 573/11/0 -- the
same numbers as before the move -- which is the proof behavior held.
The TUI keeps everything the new package doesn't need: ingest,
migration, archival, and git capture stay put.

Wire the new package into both consuming apps' dependencies and add
its typecheck to the root chain.
