+++ apps/tui/package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "dependencies": {
     "@numisma/engine": "workspace:*",
+    "@numisma/event-store": "workspace:*",
     "@opentui/core": "0.2.15"
   },
   "scripts": {
