+++ apps/tui/src/position-trimmed-reliable.test.ts
@@ -12,12 +12,12 @@ import { mkdir, mkdtemp, readFile, rm, stat, writeFile } from "node:fs/promises"
 import { resolve } from "node:path";
 import { tmpdir } from "node:os";
 import {
-  ingestInbox,
   loadEventLog,
   loadFoldedReview,
   resolveEventStorePaths,
   type EventStorePaths,
-} from "./event-store.js";
+} from "@numisma/event-store";
+import { ingestInbox } from "./event-store.js";
 import { buildCompositionReport } from "@numisma/engine";
 import { afterEach, describe, expect, it } from "vitest";
