+++ apps/web/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@numisma/engine": "workspace:*",
+    "@numisma/event-store": "workspace:*",
     "@tanstack/react-query": "5.101.2",
     "@tanstack/react-router": "1.170.17",
     "@tanstack/react-start": "1.168.27",
