+++ package.json
@@ -19,7 +19,7 @@
     "push": "tsx apps/web/src/push/push.ts",
     "db:init": "tsx apps/web/src/push/push.ts --init",
     "db:provision": "tsx apps/web/src/projection/provision-projection.ts",
-    "typecheck": "pnpm --filter @numisma/engine run typecheck && pnpm --filter @numisma/price-feed run typecheck && pnpm --filter @numisma/tui run typecheck && pnpm --filter @numisma/web run typecheck",
+    "typecheck": "pnpm --filter @numisma/engine run typecheck && pnpm --filter @numisma/event-store run typecheck && pnpm --filter @numisma/price-feed run typecheck && pnpm --filter @numisma/tui run typecheck && pnpm --filter @numisma/web run typecheck",
     "test": "vitest run",
     "coverage": "vitest run --coverage"
   },
