+++ packages/event-store/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@numisma/event-store",
+  "version": "0.7.3",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "types": "./src/index.ts",
+  "dependencies": {
+    "@numisma/engine": "workspace:*"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit -p tsconfig.json"
+  }
+}
