+++ packages/event-store/src/index.ts
@@ -1,0 +1,13 @@
+export {
+  assertLogFullyLoaded,
+  loadEventLog,
+  loadFoldedReview,
+  loadGenesis,
+  quarantineLogPath,
+  readOptional,
+  resolveDataDirDefault,
+  resolveEventStorePaths,
+  type EventLogLoad,
+  type EventStorePaths,
+  type QuarantinedLine,
+} from "./event-store.js";
